2020-11-23 2：33  重构之web参数解析重构「策略模式」

+++ Lucky/lucky-web/src/main/java/com/lucky/web/core/parameter/ParameterAnalysis.java
@@ -15,7 +15,7 @@ import java.lang.reflect.Parameter;
  */
 public interface ParameterAnalysis {
 
-    int priority();
+    double priority();
 
 
     /**
